Storing failed camera attempt

+++ src/debug/debug.c
@@ -1,7 +1,7 @@
 #include "debug.h"
 
 void printFloatArray(f32 *arr, size_t length) {
-  for (int i = 0; i < length; i++) {
+  for (u8 i = 0; i < length; i++) {
     printf("Index %d: %.5f\n", i, arr[i]);
   }
 }
